Adds stubs for additional game logic and UI change
Adds stub in the game logic for checking if the selected sector is
occupied.

Adds tweaks to test image by modifying html image to a particular
height and width.

Other changes are non consequential and were done to show Marie that
background images set in the CSS are not working