Adds core logic of occupiedSector()
Function occupiedSector() now returns false if neither x or o occupy
a given index in the game array.  Returns true if x or o do occupy
the space. A message is sent to the page telling user to try again.

Also added a new image to gameboard sector 5.  I just needed a little
win to keep the confidence up